fix circular dep check
